requirements.txt
django==3.0.3
djangorestframework==3.11.0
psycopg2-binary==2.8.4
redis==3.4.1
rq==1.2.2
python-dotenv==0.11.0
django-filter==2.2.0
